feat(mri): run-only managed transaction context, matching JRuby
execute_read / execute_write yielded the full Transaction (commit, rollback,
close) on MRI, while JRuby yields Java's DelegatingTransactionContext — a
run-only context — as Neo4j::Driver::Transaction. A managed work block must
not close the transaction itself (the driver owns commit/rollback + retry),
and downstream code prepending onto Neo4j::Driver::Transaction should target
the managed context on both flavours.

- Rename the heavy transaction class to UnmanagedTransaction (full lifecycle),
  still yielded/returned by session.begin_transaction.
- Add Neo4j::Driver::Transaction: a thin managed context exposing only #run,
  delegating to the UnmanagedTransaction the driver owns.
- Session#run_managed_transaction yields Transaction.new(@transaction);
  begin_transaction keeps yielding the UnmanagedTransaction.
- Repoint the grandfathered Metrics baseline in .rubocop_todo.yml to the
  renamed file (same code, no new offenses).

Shared spec asserts the run-only managed context and the full explicit
transaction on both flavours. Full MRI suite green (537, 1 pre-existing pending).
