+++ .rubocop_todo.yml
@@ -155,7 +155,7 @@ Metrics/AbcSize:
     - 'lib/mri/neo4j/driver/routing/routing_table.rb'
     - 'lib/mri/neo4j/driver/session.rb'
     - 'lib/mri/neo4j/driver/summary/summary_counters.rb'
-    - 'lib/mri/neo4j/driver/transaction.rb'
+    - 'lib/mri/neo4j/driver/unmanaged_transaction.rb'
     - 'lib/shared/neo4j-ruby-driver_loader.rb'
     - 'lib/shared/neo4j/driver/types/duration.rb'
     - 'spec/mri/neo4j/driver/bolt/wire_spec.rb'
@@ -234,7 +234,7 @@ Metrics/ClassLength:
     - 'lib/mri/neo4j/driver/routing/load_balancer.rb'
     - 'lib/mri/neo4j/driver/session.rb'
     - 'lib/mri/neo4j/driver/summary/result_summary.rb'
-    - 'lib/mri/neo4j/driver/transaction.rb'
+    - 'lib/mri/neo4j/driver/unmanaged_transaction.rb'
     - 'testkit-backend/command_processor.rb'
 
 # Offense count: 28
@@ -256,7 +256,7 @@ Metrics/CyclomaticComplexity:
     - 'lib/mri/neo4j/driver/routing/routing_table.rb'
     - 'lib/mri/neo4j/driver/session.rb'
     - 'lib/mri/neo4j/driver/summary/summary_counters.rb'
-    - 'lib/mri/neo4j/driver/transaction.rb'
+    - 'lib/mri/neo4j/driver/unmanaged_transaction.rb'
     - 'lib/shared/neo4j/driver/types/duration.rb'
     - 'lib/shared/neo4j/driver/types/point.rb'
     - 'testkit-backend/conversion.rb'
@@ -294,7 +294,7 @@ Metrics/MethodLength:
     - 'lib/mri/neo4j/driver/summary/result_summary.rb'
     - 'lib/mri/neo4j/driver/summary/server_info.rb'
     - 'lib/mri/neo4j/driver/summary/summary_counters.rb'
-    - 'lib/mri/neo4j/driver/transaction.rb'
+    - 'lib/mri/neo4j/driver/unmanaged_transaction.rb'
     - 'lib/shared/neo4j-ruby-driver_loader.rb'
     - 'lib/shared/neo4j/driver/auto_closeable.rb'
     - 'lib/shared/neo4j/driver/types/duration.rb'
@@ -326,7 +326,7 @@ Metrics/ParameterLists:
     - 'lib/mri/neo4j/driver/result.rb'
     - 'lib/mri/neo4j/driver/routing/load_balancer.rb'
     - 'lib/mri/neo4j/driver/routing/routing_table.rb'
-    - 'lib/mri/neo4j/driver/transaction.rb'
+    - 'lib/mri/neo4j/driver/unmanaged_transaction.rb'
     - 'lib/mri/neo4j/driver/types/relationship.rb'
     - 'lib/shared/neo4j/driver/exceptions/neo4j_exception.rb'
     - 'lib/shared/neo4j/driver/types/point.rb'
@@ -343,7 +343,7 @@ Metrics/PerceivedComplexity:
     - 'lib/mri/neo4j/driver/routing/load_balancer.rb'
     - 'lib/mri/neo4j/driver/session.rb'
     - 'lib/mri/neo4j/driver/summary/summary_counters.rb'
-    - 'lib/mri/neo4j/driver/transaction.rb'
+    - 'lib/mri/neo4j/driver/unmanaged_transaction.rb'
     - 'lib/shared/neo4j/driver/types/point.rb'
     - 'testkit-backend/conversion.rb'
     - 'testkit-backend/requests/start_test.rb'
